Test SIGKILL racing with syscall entry/exit, and avoid recording a ticks value that doesn't correspond to the recorded registers

+++ src/RecordTask.h
@@ -351,7 +351,10 @@ public:
    * system call being handled by syscall buffering.
    */
   bool running_inside_desched() const;
-  uint16_t get_ptrace_eventmsg_seccomp_data();
+  /**
+   * Returns -1 if we failed (the process unexpectedly exited).
+   */
+  int get_ptrace_eventmsg_seccomp_data();
 
   /**
    * Save tracee data to the trace.  |addr| is the address in
